requirements.txt
django>=5.1,<6.0
wagtail>=7.2,<8.0
psycopg2-binary>=2.9
gunicorn>=21.0
whitenoise>=6.6
python-dotenv>=1.0
dj-database-url>=2.0
pyyaml>=6.0
neo4j>=5.0
django-extensions>=3.2
